requirements.txt
selenium==4.9.0
load-dotenv==0.1.0